Create shared LogSection component
