+++ app/src/main/java/li/crescio/penates/diana/ui/LogSection.kt
@@ -1,0 +1,44 @@
+package li.crescio.penates.diana.ui
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun LogSection(logs: List<String>) {
+    val listState = rememberLazyListState()
+    LaunchedEffect(logs.size) {
+        if (logs.isNotEmpty()) {
+            listState.scrollToItem(logs.lastIndex)
+        }
+    }
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(120.dp)
+            .background(Color.Black)
+    ) {
+        LazyColumn(state = listState, modifier = Modifier.padding(8.dp)) {
+            items(logs) { log ->
+                Text(
+                    log,
+                    fontFamily = FontFamily.Monospace,
+                    color = Color.Green
+                )
+            }
+        }
+    }
+}
+
