+++ app/src/main/java/li/crescio/penates/diana/ui/NotesListScreen.kt
@@ -1,17 +1,12 @@
 package li.crescio.penates.diana.ui
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.res.stringResource
 import li.crescio.penates.diana.R
@@ -26,12 +21,6 @@ fun NotesListScreen(
     onAddMemo: () -> Unit,
 ) {
     Column(modifier = Modifier.fillMaxSize()) {
-        val listState = rememberLazyListState()
-        LaunchedEffect(logs.size) {
-            if (logs.isNotEmpty()) {
-                listState.scrollToItem(logs.lastIndex)
-            }
-        }
         Row(
             modifier = Modifier
                 .fillMaxWidth()
@@ -64,21 +53,6 @@ fun NotesListScreen(
             }
         }
 
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(120.dp)
-                .background(Color.Black)
-        ) {
-            LazyColumn(state = listState, modifier = Modifier.padding(8.dp)) {
-                items(logs) { log ->
-                    Text(
-                        log,
-                        fontFamily = FontFamily.Monospace,
-                        color = Color.Green
-                    )
-                }
-            }
-        }
+        LogSection(logs)
     }
 }
