+++ app/src/main/java/li/crescio/penates/diana/ui/RecorderScreen.kt
@@ -4,11 +4,7 @@ import android.Manifest
 import android.content.pm.PackageManager
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.material3.SnackbarHostState
@@ -16,9 +12,7 @@ import androidx.compose.material3.SnackbarResult
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.res.stringResource
 import androidx.core.content.ContextCompat
@@ -78,12 +72,6 @@ fun RecorderScreen(
     }
 
     Column(modifier = Modifier.fillMaxSize()) {
-        val listState = rememberLazyListState()
-        LaunchedEffect(logs.size) {
-            if (logs.isNotEmpty()) {
-                listState.scrollToItem(logs.lastIndex)
-            }
-        }
         Box(
             modifier = Modifier
                 .weight(1f)
@@ -118,21 +106,6 @@ fun RecorderScreen(
             ) { Text(stringResource(R.string.finish_recording)) }
         }
 
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(120.dp)
-                .background(Color.Black)
-        ) {
-            LazyColumn(state = listState, modifier = Modifier.padding(8.dp)) {
-                items(logs) { log ->
-                    Text(
-                        log,
-                        fontFamily = FontFamily.Monospace,
-                        color = Color.Green
-                    )
-                }
-            }
-        }
+        LogSection(logs)
     }
 }
